libs/mngr/pyproject.toml
[project]
name = "mngr"
version = "0.1.0"
description = "Library for managing AI coding agents across different hosts"
requires-python = ">=3.11"
dependencies = [
    "imbue-common",
    "cel-python>=0.1.5",
    "click>=8.0",
    "click-option-group>=0.5.6",
    "coolname>=2.2.0",
    "cryptography>=42.0",
    "modal>=0.67",
    "pydantic>=2.0",
    "psutil>=5.9",
    "pyinfra>=3.0",
    "pluggy>=1.5.0",
    "tabulate>=0.9.0",
    "urwid>=2.2.0",
]

[project.scripts]
mngr = "imbue.mngr.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
imbue-common = { workspace = true }

[tool.pyright]
venvPath = "../.."
venv = ".venv"
pythonVersion = "3.11"
strict = ["**/*.py"]
